+++ client/src/components/StepSlider.jsx
@@ -1,8 +1,15 @@
 import Slider from '@mui/material/Slider'
 import { styled } from '@mui/material/styles'
+import { useDispatch, useSelector } from 'react-redux'
 
-export const StepSlider = ({ state, dispatch }) => {
-    const percent = state.cacheStep / state.totalStep * 100
+export const StepSlider = () => {
+    const cacheStep = useSelector((state)=>state.cacheStep)
+    const totalStep = useSelector((state)=>state.totalStep)
+    const currTime = useSelector((state)=>state.currTime)
+    const sliderStep = useSelector((state)=>state.sliderStep)
+    const dispatch = useDispatch()
+
+    const percent = cacheStep / totalStep * 100
     const CustomSlider = styled(Slider)(({ theme }) => ({
         "& .MuiSlider-rail": {
             background: `linear-gradient(to right, #dedede, #dedede ${percent}%, #6c6c6c , #6c6c6c ${percent}%)`,
@@ -12,22 +19,22 @@ export const StepSlider = ({ state, dispatch }) => {
         dispatch({ type: 'PAUSE' })
         dispatch({
             type: 'SLIDER_STEP',
-            sliderStep: Math.min(newValue, state.cacheStep),
+            sliderStep: Math.min(newValue, cacheStep),
         })
-        if (performance.now() - state.currTime > 250) {
+        if (performance.now() - currTime > 250) {
             dispatch({
                 type: 'RENDER_STEP',
-                renderStep: Math.min(newValue, state.cacheStep),
+                renderStep: Math.min(newValue, cacheStep),
             })
         }
     }
     return (
         <CustomSlider
-            value={typeof state.sliderStep === 'number' ? state.sliderStep : 0}
+            value={typeof sliderStep === 'number' ? sliderStep : 0}
             onChange={handleSliderChange}
             aria-labelledby="input-slider"
             min={0}
-            max={state.totalStep}
+            max={totalStep}
         />
     )
 }