+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Change log
 
+## version 0.2.1
+- MOD render to use redux
+- ADD bag volume limit in render
+
 ## version 0.2.0
 - ADD offer accept mechanism
 - MOD rewrite player action execution